♻️(frontend) rename uploadFile.ts to index.ts
We just refactored most all of our side effects to not take a jwt as
argument but directly import `appData` instead. In doing so, we
renamed them to `index`, in line with our general policy in marsha.

uploadFile was the last side effect without an index.
